requirements/requirements.txt
alembic==1.6.5
fastapi
GeoAlchemy2==0.8.5
geopandas
geoserver-rest==2.0.0
imagecodecs==2022.8.8
numpy
pika==1.2.0
psycopg2-binary
pydantic==1.8.2
python-dotenv==0.17.1
PyYAML
rasterio==1.3a3
SQLAlchemy==1.4.41
tifffile==2021.4.8
uvicorn
netCDF4==1.6.1
h5netcdf==1.0.2
xarray
aiohttp==3.8.3